termite_fieldpack/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "termite-fieldpack"
version = "0.9.0"
description = "Offline Termux/field toolchain: CAS + SQLite/FTS + provenance + signed bundles + verify/replay."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "cryptography>=41.0.0",
  "requests>=2.31.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
]

[project.scripts]
termite = "termite.cli:main"

[tool.setuptools]
packages = ["termite"]
